backend/requirements.txt
fastapi
uvicorn
paramiko
pyyaml
pyjwt
cryptography
google-generativeai
python-dotenv
